// types.ts
export enum UserRole {
  CONSUMER = 'CONSUMER',
  MERCHANT = 'MERCHANT',
  GUEST = 'GUEST'
}

export enum AppScreen {
  LOGIN = 'LOGIN',
  HOME = 'HOME',
  AVATAR = 'AVATAR',
  MERCHANT_DASHBOARD = 'MERCHANT_DASHBOARD',
  SUPPORT = 'SUPPORT',
  HOT_PRODUCTS = 'HOT_PRODUCTS',
  COMMUNITY = 'COMMUNITY',
  LIVE_STREAM = 'LIVE_STREAM',
  PROFILE = 'PROFILE',
  PROFILE_EDIT = 'PROFILE_EDIT',
  LIVE_ROOM_CUSTOMIZE = 'LIVE_ROOM_CUSTOMIZE',
  PRODUCT_DETAIL = 'PRODUCT_DETAIL',
  CART = 'CART',
  CHECKOUT = 'CHECKOUT',
  PAYMENT = 'PAYMENT',
  ORDER_LIST = 'ORDER_LIST',
  ORDER_DETAIL = 'ORDER_DETAIL',
  ADDRESS_MANAGE = 'ADDRESS_MANAGE',
  ADDRESS_EDIT = 'ADDRESS_EDIT'
}